+++ pkg/network/anonymity/queue/queue.go
@@ -3,6 +3,7 @@ package queue
 import (
 	"context"
 	"sync"
+	"sync/atomic"
 	"time"
 
 	"github.com/number571/go-peer/pkg/client"
@@ -22,21 +23,24 @@ type sMessageQueue struct {
 	fState state.IState
 	fMutex sync.Mutex
 
-	fNetworkMask uint64
-	fMsgSettings net_message.ISettings
-
 	fSettings ISettings
 	fClient   client.IClient
 
-	fMainPool sMainPool
-	fVoidPool sVoidPool
+	fNetworkMask uint64
+	fMsgSettings net_message.ISettings
+
+	fMainPool *sMainPool
+	fVoidPool *sVoidPool
 }
 
 type sMainPool struct {
-	fQueue chan net_message.IMessage
+	fCount    int64 // atomic variable
+	fQueue    chan net_message.IMessage
+	fRawQueue chan message.IMessage
 }
 
 type sVoidPool struct {
+	fCount    int64 // atomic variable
 	fQueue    chan net_message.IMessage
 	fReceiver asymmetric.IPubKey
 }
@@ -47,11 +51,12 @@ func NewMessageQueue(pSett ISettings, pClient client.IClient) IMessageQueue {
 		fMsgSettings: net_message.NewSettings(&net_message.SSettings{}),
 		fSettings:    pSett,
 		fClient:      pClient,
-		fMainPool: sMainPool{
-			fQueue: make(chan net_message.IMessage, pSett.GetMainCapacity()),
+		fMainPool: &sMainPool{
+			fQueue:    make(chan net_message.IMessage, pSett.GetMainCapacity()),
+			fRawQueue: make(chan message.IMessage, pSett.GetMainCapacity()),
 		},
-		fVoidPool: sVoidPool{
-			fQueue:    make(chan net_message.IMessage, pSett.GetPoolCapacity()),
+		fVoidPool: &sVoidPool{
+			fQueue:    make(chan net_message.IMessage, pSett.GetVoidCapacity()),
 			fReceiver: asymmetric.NewRSAPrivKey(pClient.GetPrivKey().GetSize()).GetPubKey(),
 		},
 	}
@@ -71,25 +76,67 @@ func (p *sMessageQueue) Run(pCtx context.Context) error {
 	}
 	defer func() { _ = p.fState.Disable(nil) }()
 
+	const numProcs = 2
+	chBufErr := make(chan error, numProcs)
+
+	wg := sync.WaitGroup{}
+	wg.Add(numProcs)
+
+	go p.runVoidPoolFiller(pCtx, &wg, chBufErr)
+	go p.runMainPoolFiller(pCtx, &wg, chBufErr)
+
+	wg.Wait()
+	close(chBufErr)
+
+	errList := make([]error, 0, numProcs)
+	for err := range chBufErr {
+		errList = append(errList, err)
+	}
+	return utils.MergeErrors(errList...)
+}
+
+func (p *sMessageQueue) runVoidPoolFiller(pCtx context.Context, pWg *sync.WaitGroup, chErr chan<- error) {
+	defer pWg.Done()
 	for {
 		select {
 		case <-pCtx.Done():
-			return pCtx.Err()
+			chErr <- pCtx.Err()
+			return
 		default:
 			if err := p.fillVoidPool(pCtx); err != nil {
-				return err
+				chErr <- err
+				return
+			}
+		}
+	}
+}
+
+func (p *sMessageQueue) runMainPoolFiller(pCtx context.Context, pWg *sync.WaitGroup, chErr chan<- error) {
+	defer pWg.Done()
+	for {
+		select {
+		case <-pCtx.Done():
+			chErr <- pCtx.Err()
+			return
+		case x := <-p.fMainPool.fRawQueue:
+			if err := p.fillMainPool(pCtx, x); err != nil {
+				chErr <- err
+				return
 			}
 		}
 	}
 }
 
 func (p *sMessageQueue) WithNetworkSettings(pNetworkMask uint64, pMsgSettings net_message.ISettings) IMessageQueue {
+	// stop all pools
 	p.fMutex.Lock()
 	defer p.fMutex.Unlock()
 
+	// change net_message settings
 	p.fNetworkMask = pNetworkMask
 	p.fMsgSettings = pMsgSettings
 
+	// clear all old queue state
 	for len(p.fMainPool.fQueue) > 0 {
 		<-p.fMainPool.fQueue
 	}
@@ -101,24 +148,12 @@ func (p *sMessageQueue) WithNetworkSettings(pNetworkMask uint64, pMsgSettings ne
 }
 
 func (p *sMessageQueue) EnqueueMessage(pMsg message.IMessage) error {
-	if p.mainPoolHasLimit() {
+	incCount := atomic.AddInt64(&p.fMainPool.fCount, 1)
+	if uint64(incCount) > p.fSettings.GetMainCapacity() {
+		atomic.AddInt64(&p.fMainPool.fCount, -1)
 		return ErrQueueLimit
 	}
-
-	p.fMutex.Lock()
-	msgSettings := p.fMsgSettings
-	networkMask := p.fNetworkMask
-	p.fMutex.Unlock()
-
-	p.fMainPool.fQueue <- net_message.NewMessage(
-		msgSettings,
-		payload.NewPayload(
-			networkMask,
-			pMsg.ToBytes(),
-		),
-		p.fSettings.GetParallel(),
-	)
-
+	p.fMainPool.fRawQueue <- pMsg
 	return nil
 }
 
@@ -145,8 +180,42 @@ func (p *sMessageQueue) DequeueMessage(pCtx context.Context) net_message.IMessag
 	}
 }
 
+func (p *sMessageQueue) fillMainPool(pCtx context.Context, pMsg message.IMessage) error {
+	oldNetworkMask, oldMsgSettings := p.getNetworkSettings()
+	chNetMsg := make(chan net_message.IMessage)
+	go func() {
+		defer atomic.AddInt64(&p.fMainPool.fCount, -1)
+		chNetMsg <- net_message.NewMessage(
+			oldMsgSettings,
+			payload.NewPayload(
+				oldNetworkMask,
+				pMsg.ToBytes(),
+			),
+			p.fSettings.GetParallel(),
+		)
+	}()
+
+	select {
+	case <-pCtx.Done():
+		return pCtx.Err()
+
+	case netMsg := <-chNetMsg:
+		newNetworkMask, newMsgSettings := p.getNetworkSettings()
+		settingsChanged := newNetworkMask != oldNetworkMask ||
+			newMsgSettings.GetNetworkKey() != oldMsgSettings.GetNetworkKey() ||
+			newMsgSettings.GetWorkSizeBits() != oldMsgSettings.GetWorkSizeBits()
+
+		if !settingsChanged {
+			p.fMainPool.fQueue <- netMsg
+		}
+		return nil
+	}
+}
+
 func (p *sMessageQueue) fillVoidPool(pCtx context.Context) error {
-	if p.voidPoolHasLimit() {
+	incCount := atomic.AddInt64(&p.fVoidPool.fCount, 1)
+	if uint64(incCount) > p.fSettings.GetVoidCapacity() {
+		atomic.AddInt64(&p.fVoidPool.fCount, -1)
 		select {
 		case <-pCtx.Done():
 			return pCtx.Err()
@@ -154,11 +223,6 @@ func (p *sMessageQueue) fillVoidPool(pCtx context.Context) error {
 			return nil
 		}
 	}
-
-	p.fMutex.Lock()
-	msgSettings := p.fMsgSettings
-	networkMask := p.fNetworkMask
-	p.fMutex.Unlock()
 
 	msg, err := p.fClient.EncryptPayload(
 		p.fVoidPool.fReceiver,
@@ -168,12 +232,14 @@ func (p *sMessageQueue) fillVoidPool(pCtx context.Context) error {
 		return err
 	}
 
+	oldNetworkMask, oldMsgSettings := p.getNetworkSettings()
 	chNetMsg := make(chan net_message.IMessage)
 	go func() {
+		defer atomic.AddInt64(&p.fVoidPool.fCount, -1)
 		chNetMsg <- net_message.NewMessage(
-			msgSettings,
+			oldMsgSettings,
 			payload.NewPayload(
-				networkMask,
+				oldNetworkMask,
 				msg.ToBytes(),
 			),
 			p.fSettings.GetParallel(),
@@ -183,32 +249,23 @@ func (p *sMessageQueue) fillVoidPool(pCtx context.Context) error {
 	select {
 	case <-pCtx.Done():
 		return pCtx.Err()
-	case x := <-chNetMsg:
-		p.fMutex.Lock()
-		settingsChanged := networkMask != p.fNetworkMask ||
-			msgSettings.GetNetworkKey() != p.fMsgSettings.GetNetworkKey() ||
-			msgSettings.GetWorkSizeBits() != p.fMsgSettings.GetWorkSizeBits()
-		p.fMutex.Unlock()
+
+	case netMsg := <-chNetMsg:
+		newNetworkMask, newMsgSettings := p.getNetworkSettings()
+		settingsChanged := newNetworkMask != oldNetworkMask ||
+			newMsgSettings.GetNetworkKey() != oldMsgSettings.GetNetworkKey() ||
+			newMsgSettings.GetWorkSizeBits() != oldMsgSettings.GetWorkSizeBits()
 
 		if !settingsChanged {
-			p.fVoidPool.fQueue <- x
+			p.fVoidPool.fQueue <- netMsg
 		}
 		return nil
 	}
 }
 
-func (p *sMessageQueue) mainPoolHasLimit() bool {
+func (p *sMessageQueue) getNetworkSettings() (uint64, net_message.ISettings) {
 	p.fMutex.Lock()
 	defer p.fMutex.Unlock()
 
-	currLen := len(p.fMainPool.fQueue)
-	return uint64(currLen) >= p.fSettings.GetMainCapacity()
-}
-
-func (p *sMessageQueue) voidPoolHasLimit() bool {
-	p.fMutex.Lock()
-	defer p.fMutex.Unlock()
-
-	currLen := len(p.fVoidPool.fQueue)
-	return uint64(currLen) >= p.fSettings.GetPoolCapacity()
+	return p.fNetworkMask, p.fMsgSettings
 }