+++ pkg/network/anonymity/queue/settings.go
@@ -11,7 +11,7 @@ var (
 type SSettings sSettings
 type sSettings struct {
 	FMainCapacity uint64
-	FPoolCapacity uint64
+	FVoidCapacity uint64
 	FParallel     uint64
 	FDuration     time.Duration
 }
@@ -19,7 +19,7 @@ type sSettings struct {
 func NewSettings(pSett *SSettings) ISettings {
 	return (&sSettings{
 		FMainCapacity: pSett.FMainCapacity,
-		FPoolCapacity: pSett.FPoolCapacity,
+		FVoidCapacity: pSett.FVoidCapacity,
 		FParallel:     pSett.FParallel,
 		FDuration:     pSett.FDuration,
 	}).mustNotNull()
@@ -29,8 +29,8 @@ func (p *sSettings) mustNotNull() ISettings {
 	if p.FMainCapacity == 0 {
 		panic(`p.FMainCapacity == 0`)
 	}
-	if p.FPoolCapacity == 0 {
-		panic(`p.FPoolCapacity == 0`)
+	if p.FVoidCapacity == 0 {
+		panic(`p.FVoidCapacity == 0`)
 	}
 	if p.FParallel == 0 {
 		panic(`p.FParallel == 0`)
@@ -49,8 +49,8 @@ func (p *sSettings) GetMainCapacity() uint64 {
 	return p.FMainCapacity
 }
 
-func (p *sSettings) GetPoolCapacity() uint64 {
-	return p.FPoolCapacity
+func (p *sSettings) GetVoidCapacity() uint64 {
+	return p.FVoidCapacity
 }
 
 func (p *sSettings) GetDuration() time.Duration {