+++ pkg/network/anonymity/queue/types.go
@@ -23,7 +23,7 @@ type IMessageQueue interface {
 
 type ISettings interface {
 	GetMainCapacity() uint64
-	GetPoolCapacity() uint64
+	GetVoidCapacity() uint64
 	GetParallel() uint64
 	GetDuration() time.Duration
 }